+++ R/flights.R
@@ -8,10 +8,11 @@
 #' @format Data frame with columns
 #' \describe{
 #' \item{year,month,day}{Date of departure}
-#' \item{dep_time,arr_time}{Departure and arrival times, local tz.}
+#' \item{dep_time,arr_time}{Actual departure and arrival times, local tz.}
+#' \item{sched_dep_time,sched_arr_time}{Scheduled departure and arrival times, local tz.}
 #' \item{dep_delay,arr_delay}{Departure and arrival delays, in minutes.
 #'   Negative times represent early departures/arrivals.}
-#' \item{hour,minute}{Time of departure broken in to hour and minutes}
+#' \item{hour,minute}{Time of scheduled departure broken into hour and minutes.}
 #' \item{carrier}{Two letter carrier abbreviation. See \code{\link{airlines}}
 #'   to get name}
 #' \item{tailnum}{Plane tail number}
@@ -20,5 +21,7 @@
 #'   additional metadata.}
 #' \item{air_time}{Amount of time spent in the air}
 #' \item{distance}{Distance flown}
+#' \item{time_hour}{Scheduled date and hour of the flight as a \code{POSIXct} date.
+#'   Along with \code{origin}, can be used to join flights data to weather data.}
 #' }
 "flights"